Fix type in pair warning in RGBController_LenovoUSB.cpp

+++ Controllers/LenovoControllers/LenovoUSBController/RGBController_LenovoUSB.cpp
@@ -398,15 +398,13 @@ void RGBController_LenovoUSB::DeviceUpdateLEDs()
 
         prev_zone_id = zone_id;
 
-        curr_color_map.push_back({leds[i].value & 0xFF, colors[i]});
-
+        curr_color_map.push_back({(uint8_t)(leds[i].value & 0xFF), colors[i]});
     }
 
     if(curr_color_map.size() > 0)
     {
         controller->setZoneLeds(prev_zone_id, curr_color_map);
     }
-
 }
 
 void RGBController_LenovoUSB::DeviceUpdateMode()
